+++ server/sequelize-db/query-history.js
@@ -49,7 +49,6 @@ module.exports = function(sequelize) {
       rowCount: {
         type: DataTypes.INTEGER
       },
-      // Originally createdDate
       createdAt: {
         type: DataTypes.DATE
       }